requirements.txt
django==2.2.3
psycopg2
pytest
pytest-django
djangorestframework
